+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Improvements:
 
 * Update the internal representation of identifiers to be more compact
 * Add `RoomVersionId::version_6` and `RoomVersionId::is_version_6`
+* Add `PartialOrd` and `Ord` implementations for `RoomVersionId`
 
 # 0.16.1